fix: fixed files location
