+++ packages/client/src/components/GameInfo/Modal/Modal.scss
@@ -15,7 +15,7 @@
   transition: opacity 0.1s;
 }
 
-.modal_active {
+.gameModalUI_active {
   opacity: 1;
   pointer-events: all;
 }
