+++ packages/client/src/components/GameInfo/Modal/Modal.tsx
@@ -4,22 +4,22 @@ import ModalProps from './ModalProps'
 const Modal = ({ active, children, close, title }: ModalProps) => {
   return (
     <div
-      className={active ? 'modal modal_active' : 'modal'}
+      className={active ? 'gameModalUI gameModalUI_active' : 'gameModalUI'}
       onClick={() => {
         close()
       }}>
-      <div className="modal__content" onClick={e => e.stopPropagation()}>
-        <div className="modal__header">
+      <div className="gameModalUI__content" onClick={e => e.stopPropagation()}>
+        <div className="gameModalUI__header">
           {title}
           <button
-            className="modal__close-btn"
+            className="gameModalUI__close-btn"
             onClick={() => {
               close()
             }}>
             X
           </button>
         </div>
-        <div className="modal__child-content">{children}</div>
+        <div className="gameModalUI__child-content">{children}</div>
       </div>
     </div>
   )
